requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
astropy>=5.3.0
pandas>=2.0.0
plotly>=5.14.0
Pillow>=10.0.0
